include the necessary library

+++ libft/libc/part1/ft_bzero.c
@@ -6,13 +6,14 @@
 /*   By: bportell <bportell@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/28 19:36:13 by bportell          #+#    #+#             */
-/*   Updated: 2024/10/28 20:11:32 by bportell         ###   ########.fr       */
+/*   Updated: 2024/10/29 15:11:05 by bportell         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_bzero(void *s, unsigned int n);
+#include <string.h>
+void	ft_bzero(void *s, size_t n);
 
-void	ft_bzero(void *s, unsigned int n)
+void	ft_bzero(void *s, size_t n)
 {
 	unsigned char	*ptr;
 	unsigned int	i;
